extract setLanguageInfo action creator
